src/django/requirements.txt
djangorestframework==3.5.3
django-filter==1.0.1
celery[sqs]==4.0
boto==2.45.0
django-extensions==1.7.6
